Checkpoint 1- Pinata Game
So i added comments to the Pinata Class and made a small change to your
bat class, Just changed the name of Draw method.

Have Pause Game working.
I will need to add a visual representation to Pause Screen.
Next on my list is to finish the game over method.
Maybe Add Music when the game is over.