Pinata Class 99% completed for required things.
So i finished the Reset and Hit Method.

Working on adding music and sound effects when the pinata is hit. I will
have to learn how to do that.

Im calling the isCollision in the main already and passing it the pinata
object.

Added the game over logic.

Next for me is adding music when the pinata is hit, music to the overall
game, and also style the game over menu.
OH!!! Want to add a title screen so that why they press the start game
button, the game will start and the music will start playing. Maybe the
title screen and the game screen will have different types of music.